Move tour creation into utility file.
