+++ src/TempoLab.vue
@@ -87,8 +87,6 @@
 <script setup lang="ts">
 import { computed, onBeforeMount, onMounted, ref, type Ref, useTemplateRef, watch } from "vue";
 import { storeToRefs } from "pinia";
-import { useShepherd } from "vue-shepherd";
-import type { StepOptionsButton, Tour } from "shepherd.js";
 
 import { useTempoStore, updateStoreFromJSON, serializeTempoStore } from "@/stores/app";
 
@@ -110,89 +108,6 @@ const {
   layerControlsOpen,
 } = storeToRefs(store);
 
-const backButton: StepOptionsButton = {
-  action() { return this.back(); },
-  classes: "shepherd-button-secondary",
-  text: "Back",
-};
-
-const nextButton: StepOptionsButton = {
-  action() { return this.next(); },
-  classes: "shepherd-button-secondary",
-  text: "Next",
-};
-
-const defaultButtons: StepOptionsButton[] = [backButton, nextButton];
-
-function createIntroTour(): Tour {
-  const tour = useShepherd({
-    useModalOverlay: true,
-    defaultStepOptions: {
-      buttons: defaultButtons,
-      when: {
-        show() {
-          const currentStep = tour.getCurrentStep();
-          const currentStepElement = currentStep?.getElement();
-          const content = currentStepElement?.querySelector(".shepherd-content");
-          const footer = currentStepElement?.querySelector(".shepherd-footer");
-          const progressContainer = document.createElement("div");
-          progressContainer.classList.add("progress-container");
-          const progress = document.createElement("div");
-          progress.classList.add("progress-bar");
-          const percent = 100 * (tour.steps.indexOf(currentStep) + 1) / tour.steps.length;
-          progress.style.width = `${percent}%`;
-          progress.style.backgroundColor = accentColor.value;
-          progressContainer.appendChild(progress);
-          content?.insertBefore(progressContainer, footer);
-        },
-      },
-    },
-  });
-
-  const layersPanel = document.querySelector("#layers-panel") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: layersPanel, on: "right" }, 
-    text: "This is the layers panel!",
-  });
-
-  const openCloseLayers = layersPanel.querySelector(".open-close-container") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: openCloseLayers, on: "right" },
-    text: "The layers panel can be opened and closed",
-  });
-
-  const datasetsPanel = document.querySelector("#datasets-panel") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: datasetsPanel, on: "left" }, 
-    text: "This is the datasets panel!",
-  });
-
-  const openCloseDatasets = datasetsPanel.querySelector(".open-close-container") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: openCloseDatasets, on: "left" },
-    text: "The datasets panel can also be opened and closed",
-  });
-
-  const map = document.querySelector("canvas.maplibregl-canvas") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: map, on: "top" },
-    text: "The map displays TEMPO data and other layers!",
-  });
-
-  const timeSlider = document.querySelector(".slider-row") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: timeSlider, on: "top" },
-    text: "Control the time for the currently displayed day using the slider",
-  });
-
-  const mapControls = document.querySelector(".map-view") as HTMLElement;
-  tour.addStep({
-    attachTo: { element: mapControls, on: "top"},
-    text: "Adjust the date and timezone of the map display",
-  });
-
-  return tour;
-}
 
 const query = new URLSearchParams(window.location.search);
 debugMode.value = (query.get("debug") ?? process.env.VUE_APP_TEMPO_LAB_DEBUG)?.toLowerCase() == "true";
@@ -361,10 +276,6 @@ onMounted(() => {
   updateSizes(true, true);
   setHandleVisibility(leftHandle, layerControlsOpen.value);
   setHandleVisibility(rightHandle, datasetControlsOpen.value);
-
-  const tour = createIntroTour();
-  tour.start();
-
 });
 
 function setHandleVisibility(handle: Ref<HTMLElement | null>, visible: boolean) {
